fix: 🐛 meetingPractice

See merge request s04-final/s04p31d103!150

+++ backend/springboot/src/main/java/com/tak/soda/repository/MeetingPracticeRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 // Spring Data JPA에서 제공하는 JpaRepository interface를 상속받으면
 // @Repository 어노테이션 추가 안해도 됨
-public interface MeetingPracticeRepository extends JpaRepository<MeetingPractice, Integer> {
+public interface MeetingPracticeRepository extends JpaRepository<MeetingPractice, Long> {
     // SpringBoot JpaRepository 인터페이스가 제공하는 Entity의 기본적인 CRUD 메소드를 사용
 
     /**
